+++ index.js
@@ -1,7 +1,6 @@
 'use strict';
 
-var fs = require('fs'),
-	host = require('simples/lib/http/host'),
+var host = require('simples/lib/http/host'),
 	http = require('http'),
 	https = require('https'),
 	store = require('simples/lib/store'),
@@ -39,147 +38,6 @@ var simples = function () {
 
 	// Set current instance as the main host
 	this.hosts.main = this;
-};
-
-// Read the certificates for the HTTPS server
-simples.getCertificates = function (server, options, callback) {
-
-	var files = [],
-		result = {};
-
-	// Listener for file reading end
-	function onFileRead(error, content) {
-		if (error) {
-			server.emit('error', new Error('Can not read SSL certificates'));
-		} else {
-			setFileContent(content);
-		}
-	}
-
-	// Set the content of the current file and read the next one
-	function setFileContent(content) {
-
-		// Set the content of the file in the options object
-		result[files.shift()] = content;
-
-		// Read the next file or call the callback function
-		if (files.length) {
-			fs.readFile(result[files[0]], onFileRead);
-		} else {
-			callback(result);
-		}
-	}
-
-	// Process options members
-	Object.keys(options).forEach(function (element) {
-
-		// Get certificate file names
-		if (/^(?:cert|key|pfx)$/.test(element)) {
-			files.push(element);
-		}
-
-		// Copy options members
-		result[element] = options[element];
-	});
-
-	// Start reading files
-	if (files.length) {
-		fs.readFile(result[files[0]], onFileRead);
-	} else {
-		server.emit('error', new Error('No SSL certificates defined'));
-	}
-};
-
-// Returns the host object depending on the request
-simples.getHost = function (server, request) {
-
-	var headers = request.headers,
-		host = server.hosts.main,
-		hostname = '';
-
-	// Check if host is provided by the host header
-	if (headers.host) {
-
-		// Get the host name
-		hostname = headers.host.split(':')[0];
-
-		// Check for existing HTTP host
-		if (server.hosts[hostname]) {
-			host = server.hosts[hostname];
-		}
-	}
-
-	// Check for WebSocket host
-	if (headers.upgrade) {
-		hostname = url.parse(request.url).pathname;
-		host = host.routes.ws[hostname];
-	}
-
-	return host;
-};
-
-// Add event listeners to the internal instances
-simples.prepareServer = function (server, port, callback) {
-
-	// Listener for fatal errors
-	function onError(error) {
-		server.busy = false;
-		server.started = false;
-		server.emit('error', error);
-	}
-
-	// Listener for HTTP requests
-	function onRequest(request, response) {
-
-		var host = simples.getHost(server, request);
-
-		// Process the received request
-		utils.http.connectionListener(host, request, response);
-	}
-
-	// Listener for WebSocket requests
-	function onUpgrade(request, socket) {
-
-		var host = simples.getHost(server, request);
-
-		// Check for a defined WebSocket host
-		if (host) {
-			utils.ws.connectionListener(host, request);
-		} else {
-			socket.destroy();
-		}
-	}
-
-	// Attach the listeners for the primary HTTP(S) server instance
-	server.instance.on('release', function (callback) {
-
-		// Remove busy flag
-		server.busy = false;
-
-		// Call the callback function if it is defined
-		if (callback) {
-			callback();
-		}
-	}).on('error', onError).on('request', onRequest).on('upgrade', onUpgrade);
-
-	// Check for secondary HTTP server
-	if (server.secondary) {
-
-		// Manage the HTTP server depending on HTTPS server events
-		server.instance.on('open', function () {
-			server.secondary.listen(80);
-		}).on('close', function () {
-			server.secondary.close();
-		});
-
-		// Attach the listeners for the secondary HTTP server instance
-		server.secondary.on('error', function (error) {
-			server.instance.emit('error', error);
-		}).on('request', onRequest).on('upgrade', onUpgrade);
-	}
-
-	// Start the server
-	server.start(port, callback);
 };
 
 // Inherit from host
@@ -246,22 +104,20 @@ simples.prototype.start = function (port, callback) {
 
 	// Optional cases for port and callback
 	if (typeof port === 'number') {
-		if (this.secured) {
-			port = 443;
-		}
 		if (typeof callback !== 'function') {
 			callback = null;
 		}
 	} else if (typeof port === 'function') {
 		callback = port;
-		if (this.secured) {
-			port = 443;
-		} else {
-			port = this.port;
-		}
-	} else {
 		port = this.port;
+	} else {
 		callback = null;
+		port = this.port;
+	}
+
+	// Always set port 443 for HTTPS
+	if (this.secured) {
+		port = 443;
 	}
 
 	// If the server is busy wait for release
@@ -348,15 +204,15 @@ module.exports = function (port, options, callback) {
 
 	// Create the internal server instance
 	if (utils.isObject(options)) {
-		simples.getCertificates(server, options, function (result) {
+		utils.prepareSecuredServer(server, options, function (result) {
 			server.secured = true;
 			server.instance = https.Server(result);
 			server.secondary = http.Server();
-			simples.prepareServer(server, port, callback);
+			utils.prepareServer(server, port, callback);
 		});
 	} else {
 		server.instance = http.Server();
-		simples.prepareServer(server, port, callback);
+		utils.prepareServer(server, port, callback);
 	}
 
 	return server;